requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.1
Flask-Mail==0.10.0
Werkzeug==3.0.3
SQLAlchemy==2.0.31
PyMySQL==1.1.1
qrcode[pil]==7.4.2
Pillow==10.4.0
python-dotenv==1.0.1
APScheduler==3.10.4
